+++ src/test/java/org/jenkinsci/plugins/workflow/log/tee/TeeOutputStreamTest.java
@@ -1,68 +1,44 @@
 package org.jenkinsci.plugins.workflow.log.tee;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.emptyString;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.fail;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.file.Files;
-import org.jenkinsci.plugins.workflow.log.FileLogStorage;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Rule;
+import java.io.OutputStream;
+import java.util.logging.Logger;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
-import org.jvnet.hudson.test.JenkinsRule;
-import org.jvnet.hudson.test.LoggerRule;
 
 /**
  * Tests related to potential exceptions when using TeeLogStorage.
  */
 public class TeeOutputStreamTest {
-    @ClassRule
-    public static JenkinsRule r = new JenkinsRule();
 
-    @ClassRule
-    public static LoggerRule logging = new LoggerRule();
-
-    @Rule
-    public TemporaryFolder tmp = new TemporaryFolder();
+    private static final Logger LOGGER = Logger.getLogger(TeeOutputStreamTest.class.getName());
 
     private File fileLogStorageFileA;
     private File fileLogStorageFileB;
     private File remoteCustomFileLogStorageFile;
     private static final String CONTENT = "Hello World";
 
-    @Before
-    public void before() throws Exception {
-        fileLogStorageFileA = tmp.newFile();
-        fileLogStorageFileB = tmp.newFile();
-        remoteCustomFileLogStorageFile = tmp.newFile();
-    }
-
     /**
      * Test {@link TeeOutputStream#write(int)}
      */
     @Test
-    public void primary_fails_write_char() throws Exception {
-        char content = 'a';
-        var ls = primaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
+    public void fails_write_char() throws Exception {
+        var out = multipleFails(new BufferedOutputStream(OutputStream.nullOutputStream()) {
             @Override
             public void write(int b) throws IOException {
-                throw new IOException();
+                throw new IOException("Exception for test");
             }
         });
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().write(content);
-        } catch (IOException e) {
+        try {
+            out.write(0);
             fail();
-        } finally {
-            assertCustomFileEmpty(String.valueOf(content));
+        } catch (IOException e) {
+            assertException(e);
         }
     }
 
@@ -70,19 +46,18 @@ public class TeeOutputStreamTest {
      *Test {@link TeeOutputStream#write(byte[], int, int)}
      */
     @Test
-    public void primary_fails_write_string() throws Exception {
-        var ls = primaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
+    public void fails_write_string() throws Exception {
+        var out = multipleFails(new BufferedOutputStream(OutputStream.nullOutputStream()) {
             @Override
             public void write(byte[] b, int off, int len) throws IOException {
-                throw new IOException();
+                throw new IOException("Exception for test");
             }
         });
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
-        } catch (IOException e) {
+        try {
+            out.write(new byte[] {0}, 0, 1);
             fail();
-        } finally {
-            assertCustomFileEmpty(CONTENT);
+        } catch (IOException e) {
+            assertException(e);
         }
     }
 
@@ -90,20 +65,18 @@ public class TeeOutputStreamTest {
      * Test {@link TeeOutputStream#flush()}
      */
     @Test
-    public void primary_fails_flush() throws Exception {
-        var ls = primaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
+    public void fails_flush() throws Exception {
+        var out = multipleFails(new BufferedOutputStream(OutputStream.nullOutputStream()) {
             @Override
             public void flush() throws IOException {
                 throw new IOException("Exception for test");
             }
         });
-
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
+        try {
+            out.flush();
+            fail();
         } catch (IOException e) {
-            assertThat(e.getMessage(), is("Exception for test"));
-        } finally {
-            assertCustomFileEmpty(CONTENT);
+            assertException(e);
         }
     }
 
@@ -111,132 +84,30 @@ public class TeeOutputStreamTest {
      * Test {@link TeeOutputStream#close()}
      */
     @Test
-    @Ignore
-    public void primary_fails_close() throws Exception {
-        var ls = primaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
+    public void fails_close() throws Exception {
+        var out = multipleFails(new BufferedOutputStream(OutputStream.nullOutputStream()) {
             @Override
             public void close() throws IOException {
                 throw new IOException("Exception for test");
             }
         });
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
-        } catch (IOException e) {
-            assertThat(e.getMessage(), is("Exception for test"));
-        } finally {
-            assertCustomFileEmpty(CONTENT);
-        }
-    }
-
-    /**
-     * Test {@link TeeOutputStream#write(int)}
-     */
-    @Test
-    public void secondary_fails_write_char() throws Exception {
-        char content = 'a';
-        var ls = secondaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
-            @Override
-            public void write(int b) throws IOException {
-                throw new IOException();
-            }
-        });
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().write(content);
-        } catch (IOException e) {
+        try {
+            out.close();
             fail();
-        } finally {
-            assertCustomFileEmpty(String.valueOf(content));
-        }
-    }
-
-    /**
-     *Test {@link TeeOutputStream#write(byte[], int, int)}
-     */
-    @Test
-    public void secondary_fails_write_string() throws Exception {
-        var ls = secondaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
-            @Override
-            public void write(byte[] b, int off, int len) throws IOException {
-                throw new IOException();
-            }
-        });
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
         } catch (IOException e) {
-            fail();
-        } finally {
-            assertCustomFileEmpty(CONTENT);
+            assertException(e);
         }
     }
 
-    /**
-     * Test {@link TeeOutputStream#flush()}
-     */
-    @Test
-    public void secondary_fails_flush() throws Exception {
-        var ls = secondaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
-            @Override
-            public void flush() throws IOException {
-                throw new IOException("Exception for test");
-            }
-        });
-
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
-        } catch (IOException e) {
-            assertThat(e.getMessage(), is("Exception for test"));
-        } finally {
-            assertCustomFileEmpty(CONTENT);
-        }
+    private TeeOutputStream multipleFails(OutputStream failingOutputStream) {
+        return new TeeOutputStream(
+                failingOutputStream,
+                new OutputStream[] {failingOutputStream, OutputStream.nullOutputStream()});
     }
-
-    /**
-     * Test {@link TeeOutputStream#close()}
-     */
-    @Test
-    @Ignore
-    public void secondary_fails_close() throws Exception {
-        var ls = secondaryFails(() -> new BufferedOutputStream(new FileOutputStream(remoteCustomFileLogStorageFile)) {
-            @Override
-            public void close() throws IOException {
-                throw new IOException("Exception for test");
-            }
-        });
-
-        try (TeeBuildListener overall = (TeeBuildListener) ls.overallListener()) {
-            overall.getLogger().print(CONTENT);
-        } catch (IOException e) {
-            assertThat(e.getMessage(), is("Exception for test"));
-        } finally {
-            assertCustomFileEmpty(CONTENT);
-        }
-    }
-
-    private TeeLogStorage primaryFails(RemoteCustomFileLogStorage.OutputStreamSupplier failingOutputStream) {
-        return new TeeLogStorage(
-                RemoteCustomFileLogStorage.forFile(remoteCustomFileLogStorageFile, failingOutputStream),
-                FileLogStorage.forFile(fileLogStorageFileA),
-                FileLogStorage.forFile(fileLogStorageFileB));
-    }
-
-    private TeeLogStorage secondaryFails(RemoteCustomFileLogStorage.OutputStreamSupplier failingOutputStream) {
-        return new TeeLogStorage(
-                FileLogStorage.forFile(fileLogStorageFileA),
-                RemoteCustomFileLogStorage.forFile(remoteCustomFileLogStorageFile, failingOutputStream),
-                FileLogStorage.forFile(fileLogStorageFileB));
-    }
-
-    private void assertCustomFileEmpty(String content) throws IOException {
-        var remoteCustomFileLogStorageContent = getContent(remoteCustomFileLogStorageFile);
-        var fileLogStorageFileAContent = getContent(fileLogStorageFileA);
-        var fileLogStorageFileBContent = getContent(fileLogStorageFileB);
-
-        assertThat(remoteCustomFileLogStorageContent, emptyString());
-        assertThat(fileLogStorageFileAContent, is(content));
-        assertThat(fileLogStorageFileBContent, is(content));
-    }
-
-    private String getContent(File file) throws IOException {
-        return Files.readString(file.toPath());
+    
+    private void assertException(IOException e) {
+        assertThat(e.getMessage(), is("Exception for test"));
+        assertThat(e.getSuppressed().length, is(1));
+        assertThat(e.getSuppressed()[0].getMessage(), is("Exception for test"));
     }
 }
